requirements.txt
beautifulsoup4==4.12.2
flake8==6.1.0
numpy==1.24.4
python-decouple==3.8
python-dotenv==1.0.0
requests==2.22.0
simplejson==3.16.0
